pyproject.toml
[project]
name = "cross-personality-rewriting"
version = "0.1.0"
description = "AI Cross-Personality Rewriting: experimental framework for studying how conflict between stylistic-cognitive personas affects creativity and coherence in text rewriting."
readme = "README.md"
requires-python = ">=3.10"
license = { text = "MIT" }
authors = [{ name = "Innovation Research" }]

dependencies = [
    # API clients (install any subset you have keys for)
    "openai>=1.40.0",
    "anthropic>=0.34.0",
    "httpx>=0.27.0",
    # Config / IO
    "pyyaml>=6.0",
    "python-dotenv>=1.0.1",
    "tenacity>=8.5.0",
    "diskcache>=5.6.3",
    # Embeddings / local models (CPU-OK)
    "sentence-transformers>=3.0.1",
    "transformers>=4.44.0",
    "torch>=2.2.0",
    # Scientific stack
    "numpy>=1.26.0",
    "scipy>=1.13.0",
    "pandas>=2.2.0",
    "scikit-learn>=1.5.0",
    "statsmodels>=0.14.2",
    # Plotting
    "matplotlib>=3.8.0",
    "seaborn>=0.13.2",
    # Utilities
    "tqdm>=4.66.0",
    "rich>=13.7.0",
]

[project.optional-dependencies]
dev = ["pytest>=8.0.0", "ruff>=0.5.0"]
# TTCW external calibration (Salesforce/ttcw_creativity_eval via Hugging Face)
calibration = ["datasets>=2.19.0"]
# Source corpus bootstrap from Hugging Face datasets
corpus = ["datasets>=2.19.0"]

[build-system]
requires = ["setuptools>=68.0"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
where = ["."]
include = ["src*"]

[tool.ruff]
line-length = 110
target-version = "py310"
